Reduce memory usage and little fix

+++ App/EquipmentManager.h
@@ -6,17 +6,18 @@
 #include <DHT.h>
 #include <Adafruit_NeoPixel.h>
 #include "PinList.h"
+#include "RoutineManager.h"
 
 // 📋 Structure pour les équipements
 struct Equipment {
-    int pin; // Numéro du pin
+    byte pin; // Numéro du pin
     const char* topic; // Topic MQTT associé
-    void (*action)(int, const char*); // Fonction de gestion
+    void (*action)(short int, const char*); // Fonction de gestion
 };
 
 // 📋 Déclaration des équipements (variables globales)
 extern Equipment equipmentList[];
-extern const int numEquipments;
+extern const short int numEquipments;
 extern Servo windowServo;               
 extern Servo doorServo;
 extern LiquidCrystal_I2C lcd;
@@ -25,12 +26,12 @@ extern Adafruit_NeoPixel pixels;
 
 // 📋 Prototypes des fonctions
 void setupEquipments();
-void setDigitalOutput(int pin, const char* payload);
-void setAnalogOutput(int pin, const char* payload);
+void setDigitalOutput(short int pin, const char* payload);
+void setAnalogOutput(short int pin, const char* payload);
 void setServoPosition(Servo& servo, const char* payload);
-void setLCDContent(int, const char* payload);
+void setLCDContent(short int, const char* payload);
 const char* getLCDContent();
-void setBuzzerState(int pin, const char* payload);
-void setRGBColor(int, const char* payload);
+void setBuzzerState(short int pin, const char* payload);
+void setRGBColor(short int, const char* payload);
 
 #endif
